Update to use rx and tx queues
Work in progress!
